% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bivariate.R
\docType{methods}
\name{plot_bivariate}
\alias{plot_bivariate}
\alias{plot_bivariate,annotated_numeric,annotated_numeric-method}
\alias{plot_bivariate,annotated_numeric,annotated_factor-method}
\alias{plot_bivariate,annotated_numeric,annotated_integer-method}
\alias{plot_bivariate,annotated_integer,annotated_numeric-method}
\alias{plot_bivariate,annotated_integer,annotated_integer-method}
\alias{plot_bivariate,annotated_integer,annotated_factor-method}
\alias{plot_bivariate,annotated_factor,annotated_factor-method}
\alias{plot_bivariate,annotated_factor,annotated_numeric-method}
\alias{plot_bivariate,annotated_factor,annotated_integer-method}
\title{Plot bivariate data}
\usage{
plot_bivariate(x, y, ...)

\S4method{plot_bivariate}{annotated_numeric,annotated_numeric}(x, y,
  xlim = NULL, ylim = NULL, ...)

\S4method{plot_bivariate}{annotated_numeric,annotated_factor}(x, y, ...)

\S4method{plot_bivariate}{annotated_numeric,annotated_integer}(x, y, ...)

\S4method{plot_bivariate}{annotated_integer,annotated_numeric}(x, y, ...)

\S4method{plot_bivariate}{annotated_integer,annotated_integer}(x, y, ...)

\S4method{plot_bivariate}{annotated_integer,annotated_factor}(x, y, ...)

\S4method{plot_bivariate}{annotated_factor,annotated_factor}(x, y, ...)

\S4method{plot_bivariate}{annotated_factor,annotated_numeric}(x, y, ...)

\S4method{plot_bivariate}{annotated_factor,annotated_integer}(x, y, ...)
}
\description{
Convenience function for plotting univariate data. Method dispatch ensures
correct plotting of factor, integer, numeric, and logical vectors.
}
\examples{
annotation <- new("vector_annotation", label = "Test Label", unit = "Category")

int_ <- new("annotated_integer", .Data = (-5:4)[-3], annotation = annotation, label = "Test Label")
fac_ <- new("annotated_factor", .Data = 1:4, levels = letters[1:14], annotation = annotation, label = "Test Label")


}
