% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_univariate.R
\docType{methods}
\name{plot_univariate}
\alias{plot_univariate}
\alias{plot_univariate,annotated_numeric-method}
\alias{plot_univariate,annotated_integer-method}
\alias{plot_univariate,annotated_factor-method}
\alias{plot_univariate,annotated_logical-method}
\title{Plot univariate data}
\usage{
plot_univariate(x, ...)

\S4method{plot_univariate}{annotated_numeric}(x, ...)

\S4method{plot_univariate}{annotated_integer}(x, main = "", ...)

\S4method{plot_univariate}{annotated_factor}(x, main = "", ...)

\S4method{plot_univariate}{annotated_logical}(x, main = "", ...)
}
\description{
Convenience function for plotting univariate data. Method dispatch ensures
correct plotting of factor, integer, numeric, and logical vectors.
}
\examples{
annotation <- new("vector_annotation", label = "Test Label", unit = "Category")

int_ <- new("annotated_integer", .Data = (-5:4)[-3], annotation = annotation, label = "Test Label")
fac_ <- new("annotated_factor", .Data = 1:4, levels = letters[1:14], annotation = annotation, label = "Test Label")

plot_univariate(int_)
plot_univariate(fac_)

}
